[Feature] Batch timestamp service
    Previously, a native procedure was added to get the next timestamp
value from sequence table, now it's enhanced to get a batch of timestamp
value in one request, the usage is:

   call dbms_tso.get_timestamp(db_name, table_name, batch_size)

+++ sql/timestamp_service.h
@@ -52,7 +52,7 @@ class TimestampService {
   bool open_base_table(thr_lock_type lock_type);
 
   /* Get next timestamp value */
-  bool get_timestamp(uint64_t &ts, const uint32_t batch = 1);
+  bool get_timestamp(uint64_t &ts, const uint64_t batch = 1);
 
   bool is_initialized() { return m_initialized; }
